Pairing proof of concept and example
Adaptation of Zak's syntax for pairing to checkers and trivial
resolution example making use of it. The example simply repeats the
certificate and is bound to succeed, being little more than a smoke
test.

The folder structure becomes now debatable: currently, there is
only one FPC subfolder, whereas another one will become needed. As
for the examples themselves, they are related to resolution, but
this is not the main point. Makefiles and TJPATHs will need
extensions.